renmaed some directories
