+++ akka/src/test/scala/run/cosy/http/messages/AkkaMsgInterpreter.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package run.cosy.http.message
+package run.cosy.http.messages
 
 import run.cosy.http.Http
 import run.cosy.akka.http.AkkaTp.HT
@@ -23,20 +23,10 @@ import cats.Id
 import run.cosy.http.Http.Request
 import akka.http.scaladsl.model.HttpMethods.*
 import akka.http.scaladsl.model.headers.*
-import akka.http.scaladsl.model.{
-  ContentTypes,
-  DateTime,
-  HttpEntity,
-  HttpMethods,
-  HttpRequest,
-  HttpResponse,
-  MediaTypes,
-  StatusCodes,
-  Uri
-}
+import akka.http.scaladsl.model.{ContentTypes, DateTime, HttpEntity, HttpMethods, HttpRequest, HttpResponse, MediaTypes, StatusCodes, Uri}
+
 import scala.language.implicitConversions
-import HttpMessageDB as DB
-import run.cosy.http.message.MessageInterpreterError
+import run.cosy.http.message.{MessageInterpreterError, Platform}
 object AkkaMsgInterpreter extends run.cosy.http.message.HttpMsgInterpreter[Id, AkkaTp.HT]:
 
    override def asRequest(header: DB.RequestStr): Http.Request[Id, AkkaTp.HT] =
