+++ http4s/shared/src/test/scala/run/cosy/http4s/message/Http4sAtComponentSuite.scala
@@ -1,0 +1,12 @@
+package run.cosy.http.message
+
+import run.cosy.http.Http
+import run.cosy.http.messages.{AtComponents, ServerContext}
+import run.cosy.http4s.Http4sTp.HT
+import run.cosy.http4s.Http4sTp
+
+class AkkaAtComponentSuite[F[_]] extends AtComponentSuite[F, Http4sTp.HT]:
+
+  def at(using ServerContext): AtComponents[cats.Id, Http4sTp.HT] =
+    new run.cosy.http4s.headers
+  def interp: HttpMsgInterpreter[cats.Id, Http4sTp.HT] = Http4sMsgInterpreter[]
