+++ ietfSigTests/shared/src/main/scala/run/cosy/http/message/AtComponentSuite.scala
@@ -1,15 +1,16 @@
 package run.cosy.http.message
 
 import munit.CatsEffectSuite
-import run.cosy.http.Http.{Request,Response}
-import run.cosy.http.headers.AtSelector
+import run.cosy.http.Http.{Request, Response}
+import run.cosy.http.messages.AtSelector
+import run.cosy.http.messages.{AtComponents, ServerContext}
 import run.cosy.http.{Http, auth}
+
 import scala.util.Success
-import run.cosy.http.headers.ServerContext
 import scala.util.Try
 
 trait AtComponentSuite[F[_], H <: Http] extends CatsEffectSuite:
-   def at(using ServerContext): run.cosy.http.headers.AtComponents[F, H]
+   def at(using ServerContext): AtComponents[F, H]
    def interp: HttpMsgInterpreter[F, H]
 
    test("@method") {
@@ -252,16 +253,3 @@ trait AtComponentSuite[F[_], H <: Http] extends CatsEffectSuite:
    }
 
 end AtComponentSuite
-
-// (
-//      "@authority" -> "www.example.com",
-//      "@authority;req" -> "www.example.com",
-//      "@request-target" -> "/path?param=value",
-//      "@request-target;req" -> "/path?param=value",
-//      "@path" -> "/path",
-//      "@path;req" -> "/path",
-//      "@query" -> "/path?param=value",
-//      "@query;req" -> "/path?param=value",
-//      "@query-param;name" -> ""
-// )
-//  }
